+++ .swiftlint.yml
@@ -133,6 +133,8 @@ identifier_name:
     - top
     - now
     - row
+    - tab
+    - app
 number_separator:
   minimum_length: 5
 reporter: "xcode" # reporter type (xcode, json, csv, checkstyle, codeclimate, junit, html, emoji, sonarqube, markdown, github-actions-logging)
