+++ Shared/Extensions/String+Substring.swift
@@ -1,0 +1,32 @@
+//
+//  String+Substring.swift
+//  Planetary
+//
+//  Created by Matthew Lorentz on 7/20/22.
+//  Copyright © 2022 Verse Communications Inc. All rights reserved.
+//
+
+import Foundation
+
+extension String {
+    func indices(of occurrence: String) -> [Int] {
+        var indices = [Int]()
+        var position = startIndex
+        while let range = range(of: occurrence, range: position..<endIndex) {
+            let i = distance(from: startIndex, to: range.lowerBound)
+            indices.append(i)
+            let offset = occurrence.distance(from: occurrence.startIndex, to: occurrence.endIndex) - 1
+            guard let after = index(range.lowerBound, offsetBy: offset, limitedBy: endIndex) else {
+                break
+            }
+            position = index(after: after)
+        }
+        return indices
+    }
+    
+    func ranges(of searchString: String) -> [Range<String.Index>] {
+        let indices = indices(of: searchString)
+        let count = searchString.count
+        return indices.map({ index(startIndex, offsetBy: $0)..<index(startIndex, offsetBy: $0 + count) })
+    }
+}
